requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==3.2.16
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-countries==7.5.1
django-decouple==2.1
django-endless-pagination==2.0
django-environ==0.9.0
django-filter==22.1
django-ipware==4.0.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-safedelete==1.3.1
et-xmlfile==1.1.0
fontawesomefree==6.3.
gunicorn==20.1.0
openpyxl==3.1.1
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
waitress==2.1.2
whitenoise==6.3.0
